+++ frontend/pages/add-[asset].vue
@@ -1,5 +1,10 @@
 <script setup lang="ts">
+definePageMeta({
+  middleware: "auth",
+});
+
 const route = useRoute();
+
 const asset: string = route.params.asset as string;
 const allowableAssets = ["book", "author", "publisher", "tag",];